initialize player
